Create EDA_V1.py
This class performs a variety of functions including: summary stats, removal of outliers, determines collinearity, and others. I had tried to make it cluster using DBSCAN to check if the outliers determined by Z-Score and Mean Absolute Deviation were correct, and while it works for small datasets it is computationally expensive.